Address feedback
Move AST to a separate folder, move classes into separate files.
Use parametrized expression for documentation file.
Apply simplification for ParseRuleContent.
Remove 'and' calculation out of loop.
Simplify number construction.
