+++ src/Jeffijoe.MessageFormat.MetadataGenerator/Plural/Parsing/AST/ILeftOperand.cs
@@ -1,0 +1,6 @@
+﻿namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing.AST
+{
+    public interface ILeftOperand
+    {
+    }
+}
