+++ src/Jeffijoe.MessageFormat.MetadataGenerator/Plural/Parsing/AST/LeftOperand.cs
@@ -1,32 +1,5 @@
-﻿namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing
+﻿namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing.AST
 {
-    public interface ILeftOperand
-    {
-    }
-
-    public class VariableOperand : ILeftOperand
-    {
-        public VariableOperand(OperandSymbol operand)
-        {
-            Operand = operand;
-        }
-
-        public OperandSymbol Operand { get; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is VariableOperand op)
-                return op.Operand == Operand;
-
-            return base.Equals(obj);
-        }
-
-        public override int GetHashCode()
-        {
-            return Operand.GetHashCode();
-        }
-    }
-
     public class ModuloOperand : ILeftOperand
     {
         public ModuloOperand(OperandSymbol operandSymbol, int modValue)
