+++ src/Jeffijoe.MessageFormat.MetadataGenerator/Plural/Parsing/AST/Relation.cs
@@ -1,0 +1,7 @@
+﻿namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing.AST
+{
+    public enum Relation
+    {
+        Equals, NotEquals
+    }
+}
