--- src/Jeffijoe.MessageFormat.MetadataGenerator/Plural/Parsing/Relation.cs
@@ -1,7 +1,0 @@
-﻿namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing
-{
-    public enum Relation
-    {
-        Equals, NotEquals
-    }
-}
