+++ src/Jeffijoe.MessageFormat.MetadataGenerator/Plural/SourceGeneration/RuleGenerator.cs
@@ -1,8 +1,7 @@
-﻿using Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing;
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing.AST;
 
 namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.SourceGeneration
 {
